+++ ui/components/badges/base/example.jsx
@@ -2,8 +2,7 @@
 // Licensed under BSD 3-Clause - see LICENSE.txt or git.io/sfdc-license
 
 import React from 'react';
-import { UtilityIcon } from '../../icons/base/example';
-import { Badge, InverseBadge, LightestBadge } from '../';
+import { Badge, InverseBadge, LightestBadge, BadgeIcon } from '../';
 
 export default <Badge>Badge Label</Badge>;
 
@@ -11,21 +10,66 @@ export let examples = [
   {
     id: 'inverse',
     label: 'Inverse',
-    element: <InverseBadge>Badge Label</InverseBadge>
+    element: <InverseBadge>Inverse badge</InverseBadge>
   },
   {
-    id: 'light-with-icon',
-    label: 'Light with Icon',
+    id: 'with-left-icon',
+    label: 'With left icon',
+    element: (
+      <Badge>
+        <BadgeIcon symbol="moneybag" />
+        423 Credits Available
+      </Badge>
+    )
+  },
+  {
+    id: 'with-right-icon',
+    label: 'With right icon',
+    element: (
+      <Badge>
+        423 Credits Available
+        <BadgeIcon symbol="moneybag" align="right" />
+      </Badge>
+    )
+  },
+  {
+    id: 'light-with-left-icon',
+    label: 'Light with left icon',
     element: (
       <LightestBadge>
-        <UtilityIcon
-          containerClassName="slds-m-right_xx-small"
-          className="slds-icon_xx-small slds-icon-text-default"
-          assistiveText={false}
-          symbol="moneybag"
-        />
+        <BadgeIcon symbol="moneybag" />
         423 Credits Available
       </LightestBadge>
+    )
+  },
+  {
+    id: 'light-with-right-icon',
+    label: 'Light with right icon',
+    element: (
+      <LightestBadge>
+        423 Credits Available
+        <BadgeIcon symbol="moneybag" align="right" />
+      </LightestBadge>
+    )
+  },
+  {
+    id: 'inverse-with-left-icon',
+    label: 'Inverse with left icon',
+    element: (
+      <InverseBadge>
+        <BadgeIcon symbol="moneybag" isInverse />
+        423 Credits Available
+      </InverseBadge>
+    )
+  },
+  {
+    id: 'inverse-with-right-icon',
+    label: 'Inverse with right icon',
+    element: (
+      <InverseBadge>
+        423 Credits Available
+        <BadgeIcon symbol="moneybag" align="right" isInverse />
+      </InverseBadge>
     )
   },
   {