+++ ui/components/badges/RELEASENOTES.md
@@ -6,6 +6,11 @@
 
 ## 2.8.0
 
+### Added
+
+- Examples for all color variants with left and right aligned placement.
+- Wrapping `slds-badge__icon` element to position and color icons.
+
 ### Changed
 
 - Increased font size to 12px, applied bold font weight, and removed uppercase styling in `.slds-badge`.